# landing/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Q-Mol API",
    "description": "Molecular descriptor, similarity search, drug-likeness screen, and ADMET prediction API.\n\nGet a free key: `POST /signup`. Upgrade: https://qmol.app/checkout.html\n\nPostman collection: [qmol-postman.json](/qmol-postman.json)",
    "contact": {
      "name": "Q-Mol support",
      "email": "hi@qmol.app"
    },
    "license": {
      "name": "MIT (free tier) / Commercial (paid tiers)"
    },
    "version": "1.2.0"
  },
  "paths": {
...
    "/signup": {
      "post": {
        "summary": "Signup",
        "description": "Self-service free-tier API key. Rate-limited to 1/min per IP.\n\nSigning up with the same email twice returns the existing key.\nOptional ?ref=<code> credits the referrer with bonus quota.",
        "operationId": "signup_signup_post",
        "parameters": [
          {
            "name": "ref",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ref"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/account": {
      "delete": {
        "summary": "Account Delete",
        "description": "Permanently delete the caller's account and all associated data.\n\nIrreversible. Removes every key under the same email plus usage, audit,\nscopes, team memberships, webhooks, and referral data.",
        "operationId": "account_delete_account_delete",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/billing/checkout": {
      "post": {
        "summary": "Billing Checkout",
        "description": "Create a Stripe Checkout *subscription* session for a tier and return the\nhosted-checkout URL. Public (the buyer has no key yet — the Stripe webhook\nprovisions + emails the key on payment). Returns 503 until Stripe is\nconfigured, so the endpoint is safe to ship before keys are set.",
        "operationId": "billing_checkout_billing_checkout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutSessionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/billing/play/verify": {
      "post": {
        "summary": "Play Verify",
        "description": "Verify a Google Play subscription purchase (Android in-app billing) and\nprovision/return the API key. Returns 503 until Play credentials are set, so\nit's safe to ship before the Android app exists.",
        "operationId": "play_verify_billing_play_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayVerifyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/coupon/check": {
      "post": {
        "summary": "Coupon Check",
        "description": "Used by checkout.html before creating a Stripe session.",
        "operationId": "coupon_check_coupon_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CouponCheckIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/magic-link": {
      "post": {
        "summary": "Magic Link Request",
        "description": "Send a magic link to the email's inbox. Rate-limited to prevent spam.",
        "operationId": "magic_link_request_auth_magic_link_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MagicLinkIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/redeem": {
      "get": {
        "summary": "Magic Link Redeem",
        "operationId": "magic_link_redeem_auth_redeem_get",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckoutSessionIn": {
        "properties": {
          "tier": {
            "type": "string",
            "minLength": 1,
            "title": "Tier"
          },
          "success_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Success Url"
          },
          "cancel_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancel Url"
          }
        },
        "type": "object",
        "required": [
          "tier"
        ],
        "title": "CheckoutSessionIn"
      },
...
      "CouponCheckIn": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "tier": {
            "type": "string",
            "title": "Tier"
          }
        },
        "type": "object",
        "required": [
          "code",
          "tier"
        ],
        "title": "CouponCheckIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MagicLinkIn": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "MagicLinkIn"
      },
...
      "PlayVerifyIn": {
        "properties": {
          "product_id": {
            "type": "string",
            "minLength": 1,
            "title": "Product Id"
          },
          "purchase_token": {
            "type": "string",
            "minLength": 1,
            "title": "Purchase Token"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "purchase_token"
        ],
        "title": "PlayVerifyIn"
      },
...
      "SignupIn": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "SignupIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}